Spices repo: add makepot of spices metadata
The `applets-metadata-makepot` script creates a .pot file with the spices names and descriptions (from the metadata).

It also creates translations (.po files) of this .pot by grabbing the translations from the spices .po files

Command:
```
./applets-metadata-makepot
```

Purpose of this .pot and .po files is to show the Spices Settings Download section translated.