rancher@centurix: Fix for Cinnamon 3.8 (#1694)
(cinnamon:97022): Cjs-WARNING **: 06:56:52.885: Some code accessed the
property 'STATUS_HOMESTEAD_MISSING' on the module 'homestead'. That
property was defined with 'let' or 'const' inside the module. This was
previously supported, but is not correct according to the ES6 standard.
Any symbols to be exported from a module must be defined with 'var'. The
property access will work as previously for the time being, but please
fix your code anyway.