hideable-applets@cardsurf: Fix for Cinnamon 3.8 (#1690)
Needed to move AppletDirectory assignment under the Cinnamon < 3.8
condition, the property access will prevent the applet from loading.

(cinnamon:97022): Cjs-WARNING **: 06:04:30.811: Some code accessed the
property 'ActionType' on the module 'appletConstants'. That property was
defined with 'let' or 'const' inside the module. This was previously
supported, but is not correct according to the ES6 standard. Any symbols
to be exported from a module must be defined with 'var'. The property
access will work as previously for the time being, but please fix your
code anyway.