[SpicesUpdate@claudiux] Update de.po (#2732)
Hi @claudiux , I waited for further po-changes, but there are none... ;-)
To update this applet in a lower frequency - would it be possible to combine all po-updates at once and release it only then? But I don't know how to manage that, and I fear it creates too much extra work.