cryptocoin@guantanamoe Add rates update interval configuration
To refresh the countervalue of a ticker, calls are made every 5 seconds
to the cryptocompare API, which can lead to reaching the API rate free
tier limit.

To avoid that, a configuration field has been added to set a custom
interval and reduce the frequency of calls made to the API.
This configuration field default value is set to 5 seconds to avoid
behavior change.