[hwmonitor@sylfurd]: Update requirements for hwmonitor
GUdev was introduced in 802098c8 (via linuxmint/cinnamon-spices-applets#2983)
as a workaround for linuxmint/cinnamon-spices-applets#2801.

Without the GUdev package the applet crashes and can not be added to
the panel. The following error is shown in `~/.xsession-errors`:

```
  Cjs-Message: 20:07:01.690: JS LOG: [LookingGlass/error]
  [hwmonitor@sylfurd]: Requiring GUdev, version none: Typelib file for namespace 'GUdev' (any version) not found
  [hwmonitor@sylfurd]: Error importing applet.js from hwmonitor@sylfurd
  Cjs-Message: 20:07:01.690: JS LOG: [LookingGlass/trace]
  <----------------
  anonymous@/home/jkirk/.local/share/cinnamon/applets/hwmonitor@sylfurd/3.8/applet.js:32:7
  createExports@/usr/share/cinnamon/js/misc/fileUtils.js:210:31
  requireModule/</<@/usr/share/cinnamon/js/misc/fileUtils.js:288:25
  ---------------->
```

The documentation needs to be updated to reflect that new dependency.

To completely fix this issue a dependency check should be added.

Supports: linuxmint/cinnamon-spices-applets#3023