WindowListGroup@jake.phy@gmail.com: Remove applet
This applet hasn't been maintained for years. It has been kept for users of Cinnamon 3.0 or minor.
Mint 18 / Cinnamon 3.0 reaches end of life soon and we do have a Grouped
Windows List by default in Cinnamon for quite some time now. There's also
the option to use CobiWindowList, which is far better supported. So
there's no reason to keep this applet any longer.