removed playerctl (#3437)
Currently we use playerctl for stopping the radio and increasing/decreasing the volume. This has several disadvantages - users have to install the package and it is very difficult to trigger actions when the volume is changed or the radio player paused from outside of the applet (e.g the sound applet). This pull request removes the playerctl dependency by using the D-Bus interface instead (similar to the sound applet).