[hamster@projecthamster.wordpress.com] Fix crash: for-each-in syntax is not supported (#3502)
Under Cinnamon 4.8.6 on Linux Mint 20.1, I tried to add the Project Hamster
applet to my Panel (task bar). The applet did not load, and the following error
appeared in Looking Glass:

    error t=2021-02-01T21:11:01Z
    [hamster2@projecthamster.wordpress.com]: missing ( after for
    [hamster2@projecthamster.wordpress.com]: Error importing applet.js from hamster2@projecthamster.wordpress.com
    trace t=2021-02-01T21:11:01Z
    <----------------
    createExports@/usr/share/cinnamon/js/misc/fileUtils.js:211:59
    requireModule/</<@/usr/share/cinnamon/js/misc/fileUtils.js:291:25
    ---------------->

The error is caused by the `for each (x in obj)` syntax in `applet.js` and
`stuff.js`. Using `for (x of obj)` syntax makes the error go away.

MDN says "JavaScript 1.6's for-each-in loops are deprecated; consider using ES6
for-of instead". [1]

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Errors/For-each-in_loops_are_deprecated