Add GitHub Action "Update Translation Status Tables"
This GitHub Action updates the Translation Status Tables (see
https://github.com/linuxmint/cinnamon-spices-applets/issues/3588) after
every push to the master branch by running the `translation-tables.py`
script. The generated tables are pushed to the
`translation-status-tables` branch.

The above mentioned issue is pinned and therefore shows up at the top of
the issues page for greater visibility.