pomodoro@gregfreeman.org add time left in tooltip
This change adds the time left (minutes:seconds) at the
end of the tooltip message, thus making it easy for
anyone who does not have the pomodoro timer enabled
in the applet panel to see the time left for the current cycle.
The tooltip message will update as often as the ticker changes.

All the messages that were already present in code were reused,
so translations will not require an update.