[CassiaWindowList@klangman] Minor fixes and enhancements (#4911)
1. When a modal dialog (transient) window is open the focus tracking
would get a bit messed up. I fixed it so the focused button will
properly reflect the button associated with the window that has the
focus.

2. Enhanced the tooltip text for the hotkeys assigned to the window.
(i.e. "<Alt><Super><e>::" -> "Alt+Super+E"). I also fixed the text
when two hotkeys are specified for a single hotkey list entry.

3. Added a context menu option under the "Pin to other workspace"
menu that will pin an application to all the workspaces so you don't
have to do it one at a time.