[CassiaWindowList@klangman] A bunch of fixes and new features (#4928)
1. Fixed the "Ungroup" action when the window-list was configured as
"Grouped" so that it will actually un-group the application. Previously
this option would only work when the window-list was configured as
"pooled", "Automatic" or "one-to-one". Insert the new window-list
buttons ahead of the original grouped button in the order they were in
while grouped.

2. Fix some cases where the "one label per application pool" option
was not behaving correctly.

3. When only two workspaces exist, replace the "Pin to other workspaces"
context menu option to a toggle option for pinning to the other
workspace. The "Pin to other workspaces" sub-menu now only appears when
there are 3 or more workspaces

4. Added an option for the "Pooled"/"Automatic" configurations that
allows users to decide if they want to see thumbnails for all
applications in a pool or just the thumbnail for the one selected
window. "Show thumbnails for all windows of an application pool"

5. Added the word "Application" to the  "Un/Group windows" context menu
item to be a bit more clear

6. Make sure the last button in a "pool" is the pinned button so that
the label does not appear for pinned applications when the user selected
no labels for pinned buttons.

7. Added new mouse action options:
"Move to previous workspace"
"Move to next workspace"
"Move to previous monitor"
"Move to next monitor"
"Move titlebar on to screen"

8 The calculation for the mouse action Ctrl/Shift + Forward/Back
combinations was not matching the enumeration and therefore actions
using those combinations were not resulting in the correct actions
performed. Actions for those 4 combinations now works correctly.

9. Don't show the "Add new hotkey for..." context menu item for windows
that are only backed by a window (i.e. no .desktop file for the app).
Adding a hotkey for this type of window would not behave as expected