[CassiaWindowList@klangman] Add new buttons on the right of the pool
1. Changed the behaviour of pooled buttons so that new buttons are
added to the right, rather then the left, of the pool. This makes it
consistent with the behaviour of new buttons that are not part of a
pool. My original decision to add new pool buttons to the left was to
avoid odd animation behaviour but some animation improvements makes
this less of an issue for me. If anyone really liked the old behaviour
let me know and I can think about adding a configuration option to
control this behaviour.

2. Introduced a window-list button removal label animation (shrink the
label until it's 0 length before completely removing the button). This
particularly helps with making the animation appear smoother when closing
the trailing button of an application pool with only one label.

3. Fixed removing existing pinned buttons when toggling off the "Allow
pinning of window list buttons" configuration option.

4. Fix focus tracking while grouping and ungrouping an application when
the focus is currently with a window in the group.

5. Added a new left click action option for grouped window-list buttons
called "Restore or hold for thumbnail menu". This option will restore
the most recent window on a click, open the Thumbnail menu when the
button is held for 350ms without mouse movement or (as per prior
behaviour) start a button drag operation if the mouse is moved before
the 350ms time expires.