[CassiaWindowList@klangman] Use CSS class grouped-window-list-box
Change the Workspace BoxLayout to use the CSS class
grouped-window-list-box which is consistent with the classes used for
other element. This has the benefit of removing some padding at the
start and end of the window-list as well as reduce the padding between
buttons on the list when on a horizontal panel.

Signed-off-by: Kevin Langman <klangman@gmail.com>