[CassiaWindowList@klangman] improvements with button pinning (#5034)
1. Added a new pinning configuration option that allows the user to
force pinned buttons to be synchronized across all workspaces. This
means that any changes to the pinned buttons (adding, deleting or
rearranging) in one workspace will effect all workspaces. This allows
you to maintain a consistent set of pinned buttons, particularly useful
for the "launcher" configuration where you might like to have a
consistent set of launcher buttons for all workspaces.

2. Added a new mouse action option called "Pin to panel toggle" which
allows you to assign the "pin to panel" action to any mouse button or
Ctrl/Shift+Button action you define. This action will toggle the pinned
setting status for the panel button you clicked on.

3. If an existing "launcher" applet exists on a panel (ether a 2nd
instance of CassiaWindowList setup in launcher mode or an instance of
panel-launchers@cinnamon.org) a new context menu item called
"Pin to launcher" will appear in the content menus. Using this context
menu item will add the application of the selected window-list button to
the launcher applet.

4. If an applet instance of panel-launchers@cinnamon.org exists but the
CassiaWindowList acquired the "panellauncher" role, it will forward any
"add to panel" requests to the panel-launchers applet rather the pinning
to the CassiaWindowList. Previously the same thing was done when a
2nd instance of CassiaWindowList (setup in launcher mode) existed.
This change just extends the same behaviour to cases where a
panel-launchers instance exists.

5. Previously the context menu option to unpin a button from a panel for
both the window-list and launcher modes was the same switch style item,
but that didn't make much sense for launcher mode. So I changed it
from a toggle item to a normal item when in launcher mode and changed
the text to "Remove from..." rather then "Pin to...".

6. Some minor fixes.