[CassiaWindowList@klangman] 4 new features and some fixes (#5048)
1. Remember manual grouped/ungrouped setting:
When you set a particular application to be grouped/ungrouped against
the default behaviour set by the configuration option "Window list
behaviour style", the application specific preference will be saved
and reapplied after a reboot or cinnamon restart.

2. All Buttons special case hotkey:
Added a special case hotkey description called "All Buttons" which
will cause the modifier+1-9 hotkeys to be mapped to the first 9 windows
in the window-list. Using this with the <Super> modifier will allow you
to have Super+1-9 work the same way that the standard Grouped Window
List works. I added a default hotkey for this new "All Buttons"
description but that will only appear in new instances (existing
instances will have to add one manually). Simply add a new hotkey with
key sequence that uses your desired modifier key(s) plus the "1" key and
set the description to "All Buttons" or simply enable it if the default
one is already in the list.

3. Hotkey hint bubbles using the "`" (aka grave) key:
Added a option to show hotkey hints when using the "`" (aka grave)
key with the hotkey modifiers. For example, if you define a hotkey using
<Super>+C assigned to the Calculator window, when you hit the <Super>+`
key the window-list button for Calculator will popup a "C" bubble to
show that the <Super>+C would select this window. If you defined hotkeys
using some other modifier sequence like <Super><Alt> then using
<Super><Alt>+` would show hints for those hotkeys.

4. Left click "new window" option for Panel Launcher mode:
Added a launcher specific configuration option that allows you to launch
a new instance of an application when using the left mouse button rather
then activating an existing window. This makes the launcher mode able to
behave more like a traditional panel-laucher. This new option will open
a new window on a click or open the thumbnail menu if the button is held
for 350ms. The new options is called "New window or hold for thumbnail
menu"

5. Fixes:
a) Fixed a bug that prevented you from grouping a set of applications
windows when one of the application window-list buttons is pinned.
b) Fixed a bug with tooltips sometimes not show the correct set of
applicable hotkeys.
c) Fixed a bug in the Automatic "Window list behaviour style" where the
automatic group/ungroup option was not always disabled when manually
grouping or ungrouping an application.