[MuteToggler@jebeaudet.com] Automatically decides which 'amixer' parameter to use (Closes #3979) (#5090)
* Puts the applet name first in log entries.

* Uses 'sh' instead of 'bash'.

* Uses 'TextDecoder' instead of 'toString'.

* Tries two versions of 'amixer' commands. (Closes #3979)

On some installations it is not enough to call 'amixer'. There it an
additional parameter is needed: 'amixer -D pulse' for amixer to find the
proper channel. This PR checks, if it can find the 'Capture' channel
with 'amixer' and if not successful with 'amixer -D pulse'.

* Adds missing import.

* Removes 'TextDecoder'.

Apparently this is not needed in this applet.

* Falls back to 'amixer' if no suitable parameter is found.

As suggested by @jebeaudet in
https://github.com/linuxmint/cinnamon-spices-applets/pull/5090#discussion_r1379477152.