[CassiaWindowList@klangman] Fix hotkey-help and new window animations
1. Clear the window count bubbles before showing the hotkey-help bubbles
so that only buttons with relevant hotkeys will have any bubbles shown
when the hotkey-help hotkey is used, even when there is no hotkey-help
to show. This is less confusing for the user when hitting the help
hotkey.

2. Fixed the new window animation to only effect the window list button
if that button is going to be the button tracking the new window.

3. Update the "Advanced" config tab help text box to document the
cinnamon fix that is required in order for the "Ctrl/Shift + mouse
button actions" list to correctly display the two middle column
contents.
https://github.com/linuxmint/cinnamon/pull/11908