[CassiaWindowList@klangman] Icon Saturation Option and fixes
1. Added a new option (under the Advanced tab in config) that allows the
user to modify the colour saturation of icons. A saturation of 0% will
result in grayscale icons and a saturation of 200% will make the icons
more vivid. The custom saturation (i.e something other then 100%) can
be applied to all buttons or just some types of buttons (minimized
buttons, idle pinned buttons, buttons on other workspaces or buttons on
other monitors). This options is experimental since there are some cases
where icons are unaffected by this setting so your mileage may vary.

2. When using the new option to include windows from other workspaces
the "Move to other workspaces" context menu would show options for an
incorrect set of workspaces, I fixed that and included the text
"(this workspace)" to the menu item for the current workspace.

3. Fixed a copy&paste coding bugs when removing the applet from the
panel. The code to remove the hotkey definitions was throwing an error

4. Change the hotkey names to have "CassiaWL" rather then "hotkey" so
that it's easier to see in the .xsession-errors when cinnamon reports
hotkey redefinitions of hotkeys defined by this applet

5. Changed the default Left-click type for launcher mode to be the
"New window or hold for thumbnail menu" option. This is more like what
people expect from a panel launcher applet

6. Fixed a possible "undefined" variable error when pooling an
application

7. Fixed a number of launcher mode bugs (particularly when the applet
is on a horizontal panel).