[directory-menu@torchipeppo] v0.3.0 - Updated release fixing detected problems
The applet does not react if any Cinnamon popup menus are on,
preventing a freezing issue.
https://github.com/linuxmint/cinnamon-spices-applets/pull/5377#issuecomment-1897670714

The GTK functionality of the applet has been moved to a Python script,
like a few other applets do, allowing the menu to be shown in recent
Cinnamon versions.
Solves https://github.com/linuxmint/cinnamon-spices-applets/issues/5387

Added settings to change icon and tooltip. Besides customization,
this allows to distinguish b/w multiple instances of the applet.

Several improvements, including the removal of deprecated GTK calls
and proper alignment of the popup menu with the applet icon.

This comes with minor known issues, reported in popup-menu.py