Added ability to change brightness by scrolling on applet icon
Added ability to scroll on the applet icon to change the brightness without having to actually open it and move the sliders. (With a lot of inspiration and code taken from the built-in sound@cinnamon.org)
Haven't tested on a multi-monitor setup, but I think it would just increase or decrease the brightness of all the monitors.