[pomodoro@gregfreeman.org] Make Pomodoro completion more obvious
Extended short break dialogs to pop-up when a pomodoro has been
completed and it is time for a short break.  Combined this new
functionality with the existing options for automatically continuing
after a pomodoro and/or short break.  In the latter case, the dialog is
closed automatically.  The dialog transitions appropriately when the
timer expires.