Use standard dialog content format and only destroy dialog on applet removed from the panel
The modal is destroyed by default on close with the new implementation, though this isn't a good solution since we use the dialog more than once on the applet, so pass an argument to not destroy the dialogs on close. Also used the standard modal dialog content structure for better formatting of the text respecting the current theme.