uiHourlyForecasts.ts: Remove rest of Summary actor, and make the summary text a tooltip of the icon instead.
The summary label was disabled in c57b27467bcd3943b3 but not
completely removed, and was still throwing allocation warnings
during updates.

(cinnamon:177153): St-CRITICAL **: 16:17:23.909: st_widget_get_theme_node called on the widget [0x5a935feae560 StLabel.hourly-data ("...")] which is not in the stage.
== Stack trace for context 0x5a935e0f5130 ==
...
...
...
...

This removes the unused actor entirely, along with all references
to it, and adds a tooltip for the hourly icon that will display
the weather summary on hover.